Add remote Lock/OCC

+++ morph-core/src/main/java/intellistream/morphstream/engine/txn/scheduler/context/ds/OCCContext.java
@@ -63,7 +63,7 @@ public class OCCContext implements SchedulerContext {
     public static class RemoteObject {
         boolean successLocked;
         String value;
-        short version;
+        int version;
         short validationVersion;
         public RemoteObject() {
             this.successLocked = false;
